{"schema":"decodex/commit/1","summary":"Split provenance note type modules","authority":"manual"}
